Componente Leaderboard creado para evitar duplciación de codigo en Sonar, solventado problema con márgen superior que ocultaba el NavBar
